pyproject.toml
[tool.black]
line-length = 79

[tool.poetry]
name = "oumodulesbot"
version = "0.1.0"
description = ""
authors = ["Jerzy Kozera <jerzy.kozera@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pylru = "^1.2.0"
"discord.py" = "^1.3.2"
httpx = "^0.11.1"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.7.9"
pytest = "^5.3.5"
pytest-asyncio = "^0.10.0"
mypy = "^0.761"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
